% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_machine_init}
\alias{docker_machine_init}
\title{Connect to a docker machine}
\usage{
docker_machine_init(machine = "default")
}
\arguments{
\item{machine}{Name of a machine}
}
\description{
Initialise environment variables to connect to a docker machine
}

