% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{dockerfile}
\alias{dockerfile}
\title{Write Dockerfile}
\usage{
dockerfile(filename = NULL)
}
\arguments{
\item{filename}{filename to write to, or \code{NULL} to return the
contents, invisibly.}
}
\description{
Write a Dockerfile
}

