% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{install_script}
\alias{install_script}
\title{Install dockertest worker script}
\usage{
install_script(destination_directory, overwrite = FALSE)
}
\arguments{
\item{destination_directory}{Directory to install to}

\item{overwrite}{Overwrite existing file?}
}
\description{
Install dockertest worker script
}

