% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dockertest_remake.R
\name{prepare_remake}
\alias{prepare_remake}
\title{Prepare for docker build}
\usage{
prepare_remake(info)
}
\arguments{
\item{info}{Result of \code{project_info_remake}, which is not
exported, so that's a bit mean.}
}
\description{
Prepare for a build by writing a dockerfile and copying scripts
into the build directory.  This will eventually merge in with the
other prepare functions, but it can't easily because it follows a
slightly different pattern.
}

