% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{build}
\alias{build}
\title{Build a docker container}
\usage{
build(type = "test", prepare = TRUE, use_cache = TRUE,
  machine = "default", filename = NULL)
}
\arguments{
\item{type}{Type of container to build.  Valid options are "test",
"run" and "production" (last one only for packages).}

\item{prepare}{Rerun \code{\link{prepare}} before building the
image?}

\item{use_cache}{Set to FALSE to skip docker's cache}

\item{machine}{name of docker machine to use}

\item{filename}{Full path to dockertest filename if not in the
current directory or in the docker/ or dockertest/ directory.}
}
\description{
Build a docker container
}

