% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{build}
\alias{build}
\title{Build a docker container}
\usage{
build(type = "test", prepare = TRUE, use_cache = TRUE,
  machine = "default")
}
\arguments{
\item{type}{Type of container to build.  Valid options are "test",
"run" and "production" (last one only for packages).}

\item{prepare}{Rerun \code{\link{prepare}} before building the
image?}

\item{use_cache}{Set to FALSE to skip docker's cache}

\item{machine}{name of docker machine to use}
}
\description{
Build a docker container
}

