% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{build}
\alias{build}
\title{Build a docker container}
\usage{
build(suffix, prepare = TRUE, use_cache = TRUE)
}
\arguments{
\item{suffix}{Suffix to use;  this will be passed through to
\code{\link{prepare}} and also used to construct the tag.}

\item{prepare}{Rerun \code{\link{prepare}} before building the
image?}

\item{use_cache}{Set to FALSE to skip docker's cache}
}
\description{
Build a docker container
}

