% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{build}
\alias{build}
\title{Build a docker container}
\usage{
build(path_build = ".", path_package = NULL, use_cache = TRUE,
  prepare = TRUE)
}
\arguments{
\item{path_build}{Path to the Dockerfile}

\item{path_package}{Path to the package}

\item{use_cache}{Set to FALSE to skip docker's cache}

\item{prepare}{Rerun \code{\link{prepare}} before building the
image?}
}
\description{
Build a docker container
}

