% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{dockerfile}
\alias{dockerfile}
\title{Write Dockerfile}
\usage{
dockerfile(path_package = NULL)
}
\arguments{
\item{path_package}{Path to the package root.  We'll try and
locate it if not specificied.}
}
\description{
Write a dockerfile
}

