% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{prepare}
\alias{prepare}
\title{Prepare for docker build}
\usage{
prepare(type = "test")
}
\arguments{
\item{type}{Type of container to build.  Valid options are "test",
"run" and "production" (last one only for packages).}
}
\description{
Prepare for a build by writing a dockerfile and copying scripts
into the build directory.
}

