% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dockertest.R
\name{prepare}
\alias{prepare}
\title{Prepare for docker build}
\usage{
prepare(suffix = "-test")
}
\arguments{
\item{suffix}{Suffix used to generate the tag and directory name}
}
\description{
Prepare for a build by writing a dockerfile and copying scripts
into the build directory.
}

